+++ src/scenes/concretes/transition_level/transition_level_scene_render.py
@@ -1,6 +1,6 @@
-from src.data.game_data import GameData
+from src.data import GameData
 from src.level import ILevelManager
-from src.utils.colors import BLACK_COLOR
+from src.utils import BLACK_COLOR
 
 from ...abstractions import Render
 from ..level.level_metrics_renderer import LevelMetricsRenderer
