+++ src/utils/assets/__init__.py
@@ -1,7 +1,3 @@
-from src.utils.assets.resources.data.high_level_score import (
-    HIGH_LEVEL_SCORE_FILE,
-)
-
 from .backgrounds.menu import CHARACTER_SELECTION_MENU, MENU_BACKGROUND
 from .cinematics import FINAL_CINEMATIC_AUDIO, FINAL_CINEMATIC_VIDEO
 from .elements import (
